==> notebooks/01_data_load.ipynb <==
## Caricamento del Dataset

In questa sezione importiamo la libreria pandas e carichiamo il dataset dei miliardari.

Il file si trova nella cartella ../data/raw/ e viene letto come un DataFrame.  
Visualizziamo le prime 5 righe per avere un'anteprima della struttura dei dati.


import pandas as pd 
df = pd.read_csv("../data/raw/Billionaires Statistics Dataset.csv")
print(df['selfMade'])
df.head()

# Distribuzione ricchezza

import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.graph_objects as go

fig1 = go.Figure(data = go.Scattergeo(
    lon = df['longitude_country'],
    lat = df['latitude_country'],
    mode = 'markers',
    marker = dict(
        size = 8, 
        opacity = .7,
        color = df['finalWorth'],
        colorscale = 'Rainbow',
        colorbar = dict(title = 'Final Worth')
    )
))
fig1.update_geos(projection_type = 'natural earth')
fig1.update_layout(title = 'Distribuzione della ricchezza miliardaria sulla mappa del mondo')
fig1.write_image("../results/mappa_ricchezza.png")
fig1.show()
