==> notebooks/06b_rule_extraction_indirect.ipynb <==
# Metodo indiretto
l'obiettivo di questo notebook è di estrarre le regole in modo indiretto dal modello decision tree finale già addestrato

carichiamo il modello che avevamo salvato nella cartella ../models

import joblib
model_path = "../models/decisiontree_selfmade_gridsearch_final.joblib"
model = joblib.load(model_path)

import pandas as pd
import numpy as np
from sklearn.tree import DecisionTreeClassifier, export_text
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix
import os
import sys
sys.path.append('../Scripts')
from utility import evaluate_and_save_model

X_train = pd.read_csv("../data/splitted/X_train.csv")
X_test = pd.read_csv("../data/splitted/X_test.csv")
y_train = pd.read_csv("../data/splitted/y_train.csv").values.ravel()
y_test = pd.read_csv("../data/splitted/y_test.csv").values.ravel()

y_pred_train = model.predict(X_train)
y_pred_test = model.predict(X_test)



evaluate_and_save_model(
    model,
    "rule based indirect",
    y_train,
    y_pred_train,
    y_test,
    y_pred_test,
    "../results/classification_selfMade/rule_based/indirect",
    "../models/indirect.joblib"
)    

- Vengono estratte le regole tramite export_text passandogli il modello e il nome delle feature e successivamente salviamo le regole su un file.txt

from sklearn.tree import export_text

tree_rules = export_text(model, feature_names=list(X_train.columns))

with open("../results/classification_selfMade/rule_based/indirect/rules.txt", "w") as f:
    f.write("=== Decision Tree Rules (indiretto) ===\n\n")
    f.write(tree_rules)
