==> notebooks/10b_baseline_category.ipynb <==
# Classificazione multiclasse
viene creato un modello random che verrà usato come baseline di confronto con modelli più complessi

import pandas as pd
import numpy as np
from sklearn.dummy import DummyClassifier
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix
import os
import sys
sys.path.append('../Scripts')
from utility import evaluate_and_save_model_multiclass

X_train = pd.read_csv("../data/splitted_category/X_train.csv")
X_test = pd.read_csv("../data/splitted_category/X_test.csv")
y_train = pd.read_csv("../data/splitted_category/y_train.csv").values.ravel()
y_test = pd.read_csv("../data/splitted_category/y_test.csv").values.ravel()

dummy = DummyClassifier(strategy='stratified', random_state=42)
dummy.fit(X_train, y_train)


y_pred_train = dummy.predict(X_train)
y_pred = dummy.predict(X_test)




evaluate_and_save_model_multiclass(
    dummy,
    "baseline",
    y_train,
    y_pred_train,
    y_test,
    y_pred,
    "../results/classification_category/baseline",
    "../models/baseline_category.joblib"
)
