==> notebooks/13c_balancedRandomForest_category.ipynb <==
# Siccome anche con random forest non siamo riusciti a colmare il divario delle performance tra test set e train set a causa di classi sbilanciate e outlier, si è deciso di utilizzare questo modello che a differenza del random forest questo è progetto per gestire le classi sbilanciate.
Nel balanced random forest:
- per ogni albero viene estratto un campione bilanciato
    - si prendono tutti gli esempi della classe minoritaria
    - poi si seleziona casualmente lo stesso numero di esempi da ciascuna delle altre classsi
    - In questo modo ogni albero vede un sottoinsieme dei dati dove le classi sono bilanciate
   - si crea la foresta
    - ogni albero vota
    - e il risultato è basato sulla maggioranza 

import pandas as pd
import numpy as np
from imblearn.ensemble import BalancedRandomForestClassifier
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix, classification_report
import matplotlib.pyplot as plt
import seaborn as sns
import os
import joblib
import sys
sys.path.append('../Scripts')
from utility import evaluate_and_save_model_multiclass

X_train = pd.read_csv("../data/splitted_category/X_train.csv")
X_test = pd.read_csv("../data/splitted_category/X_test.csv")
y_train = pd.read_csv("../data/splitted_category/y_train.csv").values.ravel()
y_test = pd.read_csv("../data/splitted_category/y_test.csv").values.ravel()

#Fitting 5 folds for each of 360 candidates, totalling 1800 fits
#Migliori iperparametri trovati:
#{'max_depth': 25, 'max_features': 'log2', 'min_samples_leaf': 1, 'min_samples_split': 2, 'n_estimators': 200}

model = BalancedRandomForestClassifier(
    n_estimators=200,
    max_depth=25,
    min_samples_leaf=1,
    min_samples_split=2,
    max_features='log2',
    random_state=42
)

model.fit(X_train, y_train)

y_train_pred = model.predict(X_train)
y_test_pred = model.predict(X_test)

evaluate_and_save_model_multiclass(
    model,
    "balanced random forest",
    y_train,
    y_train_pred,
    y_test,
    y_test_pred,
    "../results/classification_category/random forest",
    "../models/balanced_randomforest_category.joblib",
    {"max_depth":25,"n_estimators":200, "max_features":"log2"}
)

# GridSearchCV
Applichiamo la ricerca per trovare i migliori parametri che sostituiremo al nostro modello per migliorare le prestazioni

from imblearn.ensemble import BalancedRandomForestClassifier
from sklearn.model_selection import GridSearchCV

brf = BalancedRandomForestClassifier(random_state=42, n_jobs=-1)

param_grid = {
    'n_estimators': [50, 100, 150, 200],
    'max_depth': [10, 15, 20, 25, None],
    'min_samples_split': [2, 5, 10],
    'min_samples_leaf': [1, 2, 4],
    'max_features': ['sqrt', 'log2']
}

# Grid Search
grid_search = GridSearchCV(
    estimator=brf,
    param_grid=param_grid,
    scoring='f1_macro',  
    cv=5,
    n_jobs=-1,
    verbose=2
)

grid_search.fit(X_train, y_train)

print("Migliori iperparametri trovati:")
print(grid_search.best_params_)
