==> notebooks/16c_ANN_category.ipynb <==
# Ensemble voting classifier Bagging + Balanced rf + ANN
In questo notebook viene costruito un classificatore ensemble basato su soft voting che aggrega le predizioni. Quello che si fa è che per ogni istanza si prendolo le probabilità previste da ciascun modello (predict_proba) e si fa la media per ogni classe. Infine si seleziona la classe con probabilità media più alta.
Vengono caricati 3 modelli precedentemente salvati con joblib:
- bagging knn
- balanced rf
- ann
- l'obiettivo è combinare modelli diversi (che sono risultati migliori rispetto ad altri finora) per migliorare la generalizzazione

import pandas as pd
import numpy as np
from sklearn.neural_network import MLPClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix
import os
import joblib
import sys
sys.path.append('../Scripts')
from utility import evaluate_and_save_model_multiclass

X_train = pd.read_csv("../data/splitted_category/X_train.csv")
X_test = pd.read_csv("../data/splitted_category/X_test.csv")
y_train = pd.read_csv("../data/splitted_category/y_train.csv").values.ravel()
y_test = pd.read_csv("../data/splitted_category/y_test.csv").values.ravel()  

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)



bagging = joblib.load("../models/bagging_model_category.joblib")
balancedrandomforest = joblib.load("../models/balanced_randomforest_category.joblib")
ann = joblib.load("../models/ann_category.joblib")

models = [ bagging, balancedrandomforest,ann]

# predizioni sul train
probas_train = np.array([model.predict_proba(X_train_scaled) for model in models])
avg_probas_train = np.mean(probas_train, axis=0)
y_pred_train = np.argmax(avg_probas_train, axis=1)

#predizioni sul test
probas_test = np.array([model.predict_proba(X_test_scaled) for model in models])
avg_probas_test = np.mean(probas_test, axis=0)
y_pred_test = np.argmax(avg_probas_test, axis=1)

evaluate_and_save_model_multiclass(
    ann,
    "Modelli  bagging ann e balanced randomforest",
    y_train,
    y_pred_train,
    y_test,
    y_pred_test,
    "../results/classification_category/voting",
    "../models/models_category.joblib"
)
