==> notebooks/17_svm_category.ipynb <==
# Classificazione multiclasse con svm
Lo scopo è valutare le prestazioni dell'svm utilizzando il kernel rbf

## Parametri utilizzati:
- kernel='rbf': consente di gestire separazioni non lineari.
- C=1.0: bilancia l'accuratezza nel training con la capacità di generalizzare.
- gamma='scale': imposta automaticamente il parametro gamma.
- class_weight='balanced gestisce lo sbilanciamento tra le classi.

import pandas as pd
import numpy as np
from sklearn.neural_network import MLPClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix,classification_report
import os
import joblib
from sklearn.svm import SVC
import sys
sys.path.append('../Scripts')
from utility import evaluate_and_save_model_multiclass

X_train = pd.read_csv("../data/splitted_category/X_train.csv")
X_test = pd.read_csv("../data/splitted_category/X_test.csv")
y_train = pd.read_csv("../data/splitted_category/y_train.csv").values.ravel()
y_test = pd.read_csv("../data/splitted_category/y_test.csv").values.ravel()  

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)
joblib.dump(scaler, "../models/scaler_svm_category.joblib")


model = SVC(kernel='rbf', C=1.0, gamma='scale',class_weight='balanced', random_state=42)
model.fit(X_train_scaled, y_train)

y_pred_train = model.predict(X_train_scaled)
y_pred_test = model.predict(X_test_scaled)

evaluate_and_save_model_multiclass(
    model,
    "SVM",
    y_train,
    y_pred_train,
    y_test,
    y_pred_test,
    "../results/classification_category/SVM",
    "../models/svm_category.joblib",
    {"kernel":"rbf","gamma":"scale","C":1.0}
)

# GridSearchCV
In questa sezione ricerchiamo i migliori parametri per il modello svc.

from sklearn.model_selection import GridSearchCV

svc = SVC(kernel="rbf", class_weight="balanced", random_state=42)

param_grid = {
    'C': [0.1, 1, 10, 100],
    'gamma': ['scale', 0.01, 0.1, 1]
}

grid_search = GridSearchCV(
    svc,
    param_grid,
    scoring='f1_weighted',
    refit='f1_weighted',
    cv=5,
    verbose=2,
    n_jobs=-1
)

grid_search.fit(X_train_scaled, y_train)

best_model = grid_search.best_estimator_
print("Migliori parametri trovati:", grid_search.best_params_)
